Tweak example code in readme for running tests
I found that the prior example led to a very noisy test report in my terminal, which
made it a chore to scroll to the output of the specific test ns I was working on. I
suspect this might be a better default for folks like me who may not be experts
in the machinery of `clojure.test` and the various ways to run the tests, etc.